docs(native): state which operators are checked and which are refused
The capability validator's arithmetic section described a backend where
no accepted operator had an exceptional case, which stopped being true
when `add`, `sub`, `mul` and `neg` became checked. It also gave "there
is no native runtime facility" as the reason `div`, `rem`, `shl` and
`shr` are refused; that facility now exists, and the real reason is
that this release stabilizes the existing subset rather than growing
it -- each of those operators needs its own failure path and its own
differential coverage.

Also drops the last two conversions left over from the `Const::Int`
payload migration, now that both sides are signed.

+++ compiler/src/nir/lower.rs
@@ -7008,7 +7008,7 @@ mod tests {
                     result,
                     kind: ValueKind::Const(Const::Int(v)),
                     ..
-                } => Some((*result, *v as i128)),
+                } => Some((*result, *v)),
                 _ => None,
             })
             .collect();
